create food event handler- currently has delete and add new food
