+++ lib/foods/requests.js
@@ -3,33 +3,18 @@ const addFood = require('./addFood')
 const APIhost = `https://evening-peak-25213.herokuapp.com`
 const sortFoods = require('../meals/sortFoods')
 
-export const addNewFood = () => {
-  event.preventDefault()
-  $('#new-name-error').hide()
-  $('#new-calories-error').hide()
-  if ($('#new-name').val() === "") {
-    $('#new-name-error').css('display', 'block')
-    if ($('#new-calories').val() === "") {
-      $('#new-calories-error').css('display', 'block')
-    }
-  }
-  else if ($('#new-calories').val() === "") {
-    $('#new-calories-error').css('display', 'block')
-  }
-  else {
-    var data = { name: $('#new-name').val(), calories: $('#new-calories').val() }
-    return fetch(APIhost + `/api/v1/foods`, {
-      method: 'post',
-      headers:
-        { 'Content-Type': 'application/json' },
-      body: JSON.stringify(data)
-    })
-      .then(response => response.json())
-      .then(newFood => addFood.prependFood(newFood))
-      .then(resetFoodForm)
-      .catch(error => console.log({ error }))
-  }
-};
+export const addNewFoodRequest = (data) => {
+  return fetch(APIhost + `/api/v1/foods`, {
+    method: 'post',
+    headers:
+      { 'Content-Type': 'application/json' },
+    body: JSON.stringify(data)
+  })
+    .then(response => response.json())
+    .then(newFood => addFood.prependFood(newFood))
+    .then(resetFoodForm)
+    .catch(error => console.log({ error }))
+}
 
 const resetFoodForm = () => {
   $(".new-food")[0].reset();
@@ -43,9 +28,7 @@ export const getFoods = () => {
     .catch(error => console.log({ error }));
 }
 
-export const deleteFood = (event) => {
-  event.preventDefault();
-  var foodId = $('.delete-food-btn').data('food-id').toString();
+export const deleteFoodRequest = (foodId) => {
   return fetch(APIhost + `/api/v1/foods` + '/' + foodId.toString(), {
     method: 'DELETE',
     headers:
