+++ lib/index.js
@@ -1,13 +1,13 @@
 // Foods
-
+const foodEventHandler = require('./foods/eventHandler')
 const foodRequests = require('./foods/requests')
 const foodFilter = require('./foods/filter')
 
 $(document).ready(() => {
   foodRequests.getFoods();
-  $('.foods').on('click', '.delete-food-btn', foodRequests.deleteFood);
+  $('.foods').on('click', '.delete-food-btn', foodEventHandler.deleteFood);
   $('.foods').on('blur', '.food', foodRequests.editFood);
-  $('#submit').on('click', foodRequests.addNewFood);
+  $('#submit').on('click', foodEventHandler.addNewFood);
   $('#foodFilter').on('keyup', foodFilter.filterFoods);
 });
